IFC Extension 22.6 update (#573)
* IFC Extension 22.6 update

* Update steelConnections dll and release notes

* Revert default references path

* update ifc extension dlls in files to install

+++ Source/Revit.IFC.Import/Data/IFCImportFile.cs
@@ -158,7 +158,7 @@ namespace Revit.IFC.Import.Data
       /// <summary>
       /// A map of all of the already created IFC entities.  This is necessary to prevent duplication and redundant work.
       /// </summary>
-      public IDictionary<int, IFCEntity> EntityMap { get; } = new Dictionary<int, IFCEntity>();
+      public IDictionary<long, IFCEntity> EntityMap { get; } = new Dictionary<long, IFCEntity>();
 
       /// <summary>
       /// A map of all of the already created transforms for IFCLocation.  This is necessary to prevent duplication and redundant work.
